% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/marking_assignments.R
\name{anonymise_submissions}
\alias{anonymise_submissions}
\title{Anonymise all submissions to a NOW dropbox}
\usage{
anonymise_submissions(dropbox_directory, newdir)
}
\arguments{
\item{dropbox_directory}{A folder with all dropbox submissions}

\item{newdir}{}
}
\description{
Anonymise all submissions to a NOW dropbox
}
