% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/marking_assignments.R
\name{get_dropbox_submissions}
\alias{get_dropbox_submissions}
\title{Get a data-frame with details of all assessment Dropbox submissions}
\usage{
get_dropbox_submissions(dropbox_directory, most_recent_submission = TRUE)
}
\arguments{
\item{dropbox_directory}{The file path to the directory with all the
submissions to a assessment Dropbox}

\item{most_recent_submission}{Logical. Should we return only the most recent
submission by each student (default) or all submissions?}
}
\value{
A data-frame with one row per each submission, and columns for
student ID, filename, md5sum hash, and filename extension.
}
\description{
Assuming that all the contents from NOW's dropbox have been saved in a
directory, then return a data frame with the submitting student's ID, the
filename and md5sum hash and filename extension of the most recent submission
by each student, or optionally of all submissions by each subject.
}
