% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/marking_assignments.R
\name{get_marking_assignments}
\alias{get_marking_assignments}
\title{Get the marking assignments from a marking assignments directory}
\usage{
get_marking_assignments(path)
}
\arguments{
\item{path}{Path to marking assignments dir. We assume that subfolders have
the name of the marker, and inside them are the names of the marksheets.}
}
\value{
A tibble with marker name and assigned marksheets
}
\description{
Get the marking assignments from a marking assignments directory
}
