% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/marking_assignments.R
\name{get_rm2_marking_staff}
\alias{get_rm2_marking_staff}
\title{For each RM2 seminar group, get the markers and their emails}
\usage{
get_rm2_marking_staff(staff_grid_fname, staff_emails)
}
\arguments{
\item{staff_grid_fname}{A tsv file with staff_1, staff_2, seminar name, HPL}

\item{staff_emails}{A tsv file with staff name and email}
}
\value{
A long data frame
}
\description{
For each RM2 seminar group, get the markers and their emails
}
