% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/marking_assignments.R
\name{get_seminar_list}
\alias{get_seminar_list}
\title{Get classlist of all students and their labs (seminars)}
\usage{
get_seminar_list(all_seminars_fname)
}
\arguments{
\item{all_seminars_fname}{A tsv file with seminar list of labs}
}
\value{
A data frame with id, student, seminar (1_SH_LAB, etc)
}
\description{
Get classlist of all students and their labs (seminars)
}
