% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/marking_assignments.R
\name{hpl_assignments}
\alias{hpl_assignments}
\title{Assign marking assignments to HPLs}
\usage{
hpl_assignments(submissions, K = 15)
}
\arguments{
\item{submissions}{A data-frame of submissions

For each seminar}
}
\description{
Assign marking assignments to HPLs
}
