% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/marking_assignments.R
\name{marking_assignments_check}
\alias{marking_assignments_check}
\title{Check marking assignments}
\usage{
marking_assignments_check(submissions, marking_assignments)
}
\arguments{
\item{submissions}{Data-frame of reports to be marked}

\item{marking_assignments}{Data-frame of marking assignments}
}
\value{
NULL If all checks pass. Stop with error otherwise.
}
\description{
Some simple sanity checks of marking assignments. For example, each student
that submitted a report has had their report assigned to one marker, and only
one marker.
}
