% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moderation.R
\name{moderation_analysis}
\alias{moderation_analysis}
\title{Modelling Inter-marker Grading Variability}
\usage{
moderation_analysis(formula, data, ...)
}
\arguments{
\item{formula}{A formula like grade ~ marker. The lhs is the integer value of
an ordinal grade (e.g. 1 represents A+, 2 represents A- ), and the rhs is
an integer representing a unique grader.}

\item{data}{The data frame with the required variables}

\item{...}{Optional extra parameters for rstan, e.g. cores = 4.}
}
\value{
A rstan model
}
\description{
Implements a Bayesian multilevel ordinal logistic model of grading variability.
}
\examples{
\dontrun{
M <- moderation_analysis(grade ~ marker, data = exam_df)
}
}
