% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/marking_assignments.R
\name{nonhpl_marking_assignments}
\alias{nonhpl_marking_assignments}
\title{Marking assignments}
\usage{
nonhpl_marking_assignments(submissions, markers, use_hpls = FALSE)
}
\arguments{
\item{submissions}{Data frame of reports to be marked}

\item{markers}{Data frame of markers}
}
\value{
A data frame with marking assignments
}
\description{
Marking assignments
}
