% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ntupsyassessr-package.R
\docType{package}
\name{ntupsyassessr-package}
\alias{ntupsyassessr}
\alias{ntupsyassessr-package}
\title{ntupsyassessr: Tools for Assessment Tasks in the NTU Psychology Department}
\description{
Various tools for doing tasks related to assessment.
}
\author{
\strong{Maintainer}: Mark Andrews \email{mark.andrews@ntu.ac.uk}

}
\keyword{internal}
