% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/marking_assignments.R
\name{read_exported_now_mcq}
\alias{read_exported_now_mcq}
\title{Read in Exported NOW MCQ Exam CSV files}
\usage{
read_exported_now_mcq(data_dir)
}
\arguments{
\item{dat_dir}{Path name of a directory with NOW MCQ exported csv files}
}
\description{
Read in Exported NOW MCQ Exam CSV files
}
