% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/irt.R
\name{stan_irt}
\alias{stan_irt}
\title{Bayesian Item Response Theory Model}
\usage{
stan_irt(formula, data, c = NULL, ...)
}
\arguments{
\item{formula}{A formula like score ~ person + item. The lhs is the binary
score or accuracy variable. The rhs gives the person
(student/respondent/subject/participant) and item (question) variables, in
that order.}

\item{data}{The data frame with the required variables}

\item{c}{Optional guessing parameter. 0 < c < 1.}

\item{...}{Optional extra parameters for rstan, e.g. cores = 4.}
}
\value{
An rstan model
}
\description{
Implements a Bayesian multilevel 2PL IRT model
}
\examples{
\dontrun{
M <- stan_irt(score ~ nid + item, data = exam_df)
}
}
