% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/marking_assignments.R
\name{get_completed_marksheets}
\alias{get_completed_marksheets}
\title{Get list of marksheets from a directory}
\usage{
get_completed_marksheets(path)
}
\arguments{
\item{path}{The path of the directory from which to extract marksheets}
}
\value{
A tibble with one column named \code{marksheet} with marksheet files
}
\description{
Get list of marksheets from a directory
}
